requirements.txt
numpy
jax
jaxlib
flax
opencv-python
Pillow
tensorboard
tensorflow
gin-config
